% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANSI.R
\name{ANSI}
\alias{ANSI}
\title{ANSI R6 class}
\description{
ANSI R6 class

ANSI R6 class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{width, height}}{dimensions}

\item{\code{bg, fg, chr}}{matrices representing background, foreground and characters}

\item{\code{font_aspect}}{font aspect ratio}

\item{\code{bg_col, text_colour}}{initial colours}

\item{\code{ansi_bits}}{bit depth for ANSI representation.}

\item{\code{max_height}}{maximum height value which will plotted given the current
font_aspect}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{width, height}}{dimensions}

\item{\code{bg, fg, chr}}{matrices representing background, foreground and characters}

\item{\code{bg_col, text_colour}}{initial colours}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ANSI$new()}}
\item \href{#method-reset}{\code{ANSI$reset()}}
\item \href{#method-set_pixels}{\code{ANSI$set_pixels()}}
\item \href{#method-line}{\code{ANSI$line()}}
\item \href{#method-rect_filled}{\code{ANSI$rect_filled()}}
\item \href{#method-rect_outline}{\code{ANSI$rect_outline()}}
\item \href{#method-polyline}{\code{ANSI$polyline()}}
\item \href{#method-polygon}{\code{ANSI$polygon()}}
\item \href{#method-rect}{\code{ANSI$rect()}}
\item \href{#method-circle}{\code{ANSI$circle()}}
\item \href{#method-text}{\code{ANSI$text()}}
\item \href{#method-as_character_ascii}{\code{ANSI$as_character_ascii()}}
\item \href{#method-as_character}{\code{ANSI$as_character()}}
\item \href{#method-save}{\code{ANSI$save()}}
\item \href{#method-print}{\code{ANSI$print()}}
\item \href{#method-clone}{\code{ANSI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$new(
  width = 20,
  height = 20,
  background = "lightblue",
  text_colour = "black",
  ansi_bits = 8,
  font_aspect = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width, height}}{dimensions}

\item{\code{background, text_colour}}{initial colours}

\item{\code{ansi_bits}}{bit depth for ANSI representation. Possible values: 8, 24. Default:8}

\item{\code{font_aspect}}{font aspect ratio. default: 1
Reset to a blank canvas}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$reset(background = NULL, text_colour = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{background, text_colour}}{defaut bg/fg colours. default: NULL (use the colour set
       when initialised)
Set pixels in the bg}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_pixels"></a>}}
\if{latex}{\out{\hypertarget{method-set_pixels}{}}}
\subsection{Method \code{set_pixels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$set_pixels(xs, ys, bg = NULL, fg = NULL, chr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xs, ys}}{vectors of integer coordinates}

\item{\code{bg, fg, chr}}{values to set.
Draw a line}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-line"></a>}}
\if{latex}{\out{\hypertarget{method-line}{}}}
\subsection{Method \code{line()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$line(x1, y1, x2, y2, colour)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x1, y1, x2, y2}}{line extents}

\item{\code{colour}}{colour to set
Draw a filled rectangle}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rect_filled"></a>}}
\if{latex}{\out{\hypertarget{method-rect_filled}{}}}
\subsection{Method \code{rect_filled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$rect_filled(x1, y1, x2, y2, colour)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x1, y1, x2, y2}}{rect extents}

\item{\code{colour}}{colour to set
Draw a rectangular outline}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rect_outline"></a>}}
\if{latex}{\out{\hypertarget{method-rect_outline}{}}}
\subsection{Method \code{rect_outline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$rect_outline(x1, y1, x2, y2, colour)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x1, y1, x2, y2}}{rect extents}

\item{\code{colour}}{colour to set
Draw a polyline}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-polyline"></a>}}
\if{latex}{\out{\hypertarget{method-polyline}{}}}
\subsection{Method \code{polyline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$polyline(xs, ys, colour)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xs, ys}}{points around polygon}

\item{\code{colour}}{colour
Draw a polygon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-polygon"></a>}}
\if{latex}{\out{\hypertarget{method-polygon}{}}}
\subsection{Method \code{polygon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$polygon(xs, ys, colour, fill)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xs, ys}}{points around polygon}

\item{\code{colour, fill}}{colour for outline an fill
Draw a stroked/filled rectangle}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rect"></a>}}
\if{latex}{\out{\hypertarget{method-rect}{}}}
\subsection{Method \code{rect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$rect(x1, y1, x2, y2, colour, fill)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x1, y1, x2, y2}}{rect extents}

\item{\code{colour}}{colour to set}

\item{\code{fill}}{colour to set
Draw a stroked/filled rectangle

Reference: \url{http://members.chello.at/~easyfilter/bresenham.html}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-circle"></a>}}
\if{latex}{\out{\hypertarget{method-circle}{}}}
\subsection{Method \code{circle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$circle(xc, yc, r, colour, fill, font_aspect = 0.45)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xc, yc}}{circle centre}

\item{\code{r}}{radius}

\item{\code{colour}}{colour to set}

\item{\code{fill}}{colour to set}

\item{\code{font_aspect}}{font_aspect override to have circles plot as circles
Add text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-text"></a>}}
\if{latex}{\out{\hypertarget{method-text}{}}}
\subsection{Method \code{text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$text(x, y, text, colour = "black", angle = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, y}}{location}

\item{\code{text}}{string}

\item{\code{colour}}{Text colour
Character representation in ASCII onlyl}

\item{\code{angle}}{currently only supports 0 or 90.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_character_ascii"></a>}}
\if{latex}{\out{\hypertarget{method-as_character_ascii}{}}}
\subsection{Method \code{as_character_ascii()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$as_character_ascii(pow = 1, char_lookup_table = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pow}}{factor for non-linear intensity mapping}

\item{\code{char_lookup_table}}{character table. default: 1. Choices: 1, 2, 3}

\item{\code{...}}{ignored
Character representation with ANSI colouring}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_character"></a>}}
\if{latex}{\out{\hypertarget{method-as_character}{}}}
\subsection{Method \code{as_character()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$as_character(plain_ascii = FALSE, pow = 1, char_lookup_table = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plain_ascii}}{Map colour intensity to ASCII characters only. No
ANSI output.  Default: FALSE}

\item{\code{pow}}{factor for non-linear intensity mapping. Only used if
\code{plain_ascii=TRUE}}

\item{\code{char_lookup_table}}{character table. default: 1. Choices: 1, 2, 3 Only used if
\code{plain_ascii=TRUE}}

\item{\code{...}}{ignored
Save to file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$save(filename, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{filename}

\item{\code{...}}{ignored
Output to screen}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANSI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
