% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANSI.R
\name{as.character.ANSI}
\alias{as.character.ANSI}
\title{Character representation of an ANSI object}
\usage{
\method{as.character}{ANSI}(x, ...)
}
\arguments{
\item{x}{ANSI object}

\item{...}{other arguments passed on}
}
\description{
Character representation of an ANSI object
}
