% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chars.R
\name{col2char}
\alias{col2char}
\title{Colour to character lookup}
\usage{
col2char(rcolour, pow = 1, char_lookup_table = 1)
}
\arguments{
\item{rcolour}{vector of R colours e.g. 'red', '#123456'}

\item{pow}{factor for skewing the colour conversion processes}

\item{char_lookup_table}{character table. default: 1. Choices: 1, 2, 3}
}
\value{
character string. 1 character for each colour to match the intensity
        of the given colour
}
\description{
Colour to character lookup
}
