% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANSI.R
\name{collapse_matrix_to_string}
\alias{collapse_matrix_to_string}
\title{Collapse a character matrix to a string representation}
\usage{
collapse_matrix_to_string(mat, row_end = "")
}
\arguments{
\item{mat}{matrix}

\item{row_end}{character with which to end each row}
}
\value{
single string
}
\description{
Collapse a character matrix to a string representation
}
