% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chars.R
\name{colmat2char}
\alias{colmat2char}
\title{Colour matrix to character conversion}
\usage{
colmat2char(mat, pow = 1, char_lookup_table = 1)
}
\arguments{
\item{mat}{matrix of R colours e.g. 'red', '#123456'}

\item{pow}{factor for skewing the colour conversion processes}

\item{char_lookup_table}{character table. default: 1. Choices: 1, 2, 3}
}
\value{
Return a matrix as the same size of the input but replaced by a single
        character for each colour representing the grey level intensity
}
\description{
Colour matrix to character conversion
}
