% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANSI.R
\name{convert_R_colour_to_internal_colour}
\alias{convert_R_colour_to_internal_colour}
\title{Convert an R colour to an internal colour representation.}
\usage{
convert_R_colour_to_internal_colour(colour)
}
\arguments{
\item{colour}{string e.g. 'red'. '#234567'}
}
\value{
internal representation of colour
}
\description{
Currently hex colour with alpha
}
