% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon-scanline-fill.R
\name{fill_polygon}
\alias{fill_polygon}
\title{Fill a polygon}
\usage{
fill_polygon(self, xs, ys, fill, debug = FALSE)
}
\arguments{
\item{self}{R6 object with a \code{$set_pixels()} method. May be NULL}

\item{xs, ys}{coordinates of vertices (integers)}

\item{fill}{fill value. usually a character.}

\item{debug}{debug mode. default: FALSE}
}
\description{
Fill a polygon
}
