% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon-scanline-fill.R
\name{make_edges}
\alias{make_edges}
\title{Make edge information structure from polygon vertex coords}
\usage{
make_edges(xs, ys)
}
\arguments{
\item{xs, ys}{vertex coords}
}
\value{
data.frame
}
\description{
Make edge information structure from polygon vertex coords
}
